requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiomysql==0.2.0
python-dotenv==1.0.1
google-generativeai==0.4.0
aiofiles==23.2.1
